Crate threads_pool

Modules

index_mode

prelude

shared_mode

Macros

pin_mut

Structs

Config

FutPool

StatusBehaviors

ThreadPool
The standalone thread pool, which gives users more controls on the pool and where the hosted pool shall live.

Enums

ExecutionError
Enumeration to indicate possible reasons a job execution request is rejected. User will need to resubmit the job again, since closure’s state may have been stale at the execution error.

TimeoutPolicy

Traits

ConfigStatus

Hibernation

PoolManager

PoolState

StatusBehaviorSetter

ThreadPoolStates

Functions

block_on