Struct threads_pool::ThreadPool

pub struct ThreadPool { /* fields omitted */ }

The standalone thread pool, which gives users more controls on the pool and where the hosted pool shall live.

Examples

extern crate threads_pool;

use threads_pool as pool;
use std::thread;
use std::time::Duration;

// lazy-create the pool
let mut pool = pool::ThreadPool::build(4);

// get ready to submit parallel jobs, first of all, activate the pool. Not that this step is
// optional if the caller will only use the `exec` API to submit jobs.
pool.execute(|| {
    // this closure could cause panic if the pool has not been activated yet.
    println!("Executing jobs ... ");
});

// do parallel stuff.
for id in 0..10 {
    // even `id` jobs always get prioritized
    let priority = if id % 2 == 0 { true } else { false };

    // API `exec` can set if a job needs to be prioritized for execution or not.
    pool.exec(|| {
        thread::sleep(Duration::from_secs(1));
        println!("thread {} has slept for around 1 second ... ", id);
    }, priority);
}

Implementations

impl ThreadPool

pub fn new(size: usize) -> ThreadPool

Create a ThreadPool with default configurations

pub fn new_with_config(size: usize, config: Config) -> ThreadPool

Create a ThreadPool with supplied configurations

pub fn build(size: usize) -> ThreadPool

Create the ThreadPool with default pool configuration settings. When ready to activate the pool, invoke the activate_pool API before submitting jobs for execution. You can also call exec API to automatically activate the pool, however, calling the alternative immutable API execute will always lead to an error.

pub fn build_with_config(size: usize, config: Config) -> ThreadPool

Create the ThreadPool with provided pool configuration settings. When ready to activate the pool, invoke the activate_pool API before submitting jobs for execution. You can also call exec API to automatically activate the pool, however, calling the alternative immutable API execute will always lead to an error.

pub fn activate(&mut self) -> &mut Self

If the pool is lazy created, user is responsible for activating the pool before submitting jobs for execution. API exec will initialized the pool since it takes the mutable self, and hence is able to initialize the pool. However, fail to explicitly initialize the pool could cause unidentified behaviors.

Examples

extern crate threads_pool;
use std::thread;
use std::time::Duration;
use threads_pool as pool;

// lazy-create the pool
let mut pool = pool::ThreadPool::build(4);

// work on other stuff ...
thread::sleep(Duration::from_secs(4));

// get ready to submit parallel jobs, first of all, activate the pool. Not that this step is
// optional if the caller will only use the `exec` API to submit jobs.
pool.activate()
    .execute(|| {
        // this closure could cause panic if the pool has not been activated yet.
        println!("Lazy created pool now accepting new jobs...");
    });

// do parallel stuff.
for id in 0..10 {
    // API `exec` can also activate the pool automatically if it's lazy-created.
    pool.exec(|| {
        thread::sleep(Duration::from_secs(1));
        println!("thread {} has been waken after 1 seconds ... ", id);
    }, true);
}

pub fn set_exec_timeout(&mut self, timeout: Option<Duration>)

Set the time out period for a job to be queued. If timeout is defined as some duration, we will keep the new jobs in the queue for at least them amount of time when all workers of the pool are busy. If it’s set to None and all workers are busy at the moment a job comes, we will either block the caller when pool’s non_blocking setting is turned on, or return timeout error immediately when the setting is turned off.

pub fn toggle_blocking(&mut self, non_blocking: bool)

Toggle if we should block the execution APIs when all workers are busy and we can’t expand the thread pool anymore for any reasons.

pub fn toggle_auto_scale(&mut self, auto_scale: bool)

Toggle if we can add temporary workers when the ThreadPool is under pressure. The temporary works will retire after they have been idle for a period of time.

pub fn exec<F: FnOnce() + Send + 'static>(
&mut self,
f: F,
prioritized: bool
) -> Result<(), ExecutionError>

exec will dispatch a closure to a free thread to be executed. If no thread are free at the moment and the pool setting allow adding new workers at pressure, some temporary workers will be created and added to the pool for executing the accumulated pending jobs; otherwise, this API will work the same way as the alternative: execute, that it will block the caller until a worker becomes available, or the job queue timed out, where the job will be dropped and the caller needs to send the job to queue for execution again, if it’s needed.

For highly competitive environments, such as using the pool to enable async operations for a web server, disabling auto_scale could cause starvation deadlock. However, a server can only have a limited resources for disposal, we can’t create unlimited workers, there will be a limit where we have to either choose to drop the job, or put it into a queue for later execution, though that will mostly lie on caller’s discretion.

In addition, this API will take a prioritized parameter, which will allow more urgent job to be taken by the workers sooner than jobs without a priority.

Note that if the ThreadPool is created using delayed pool initialization, i.e. created using either build or build_with_config APIs, then the pool will be initialized at the first time a job is to be executed.

Examples

extern crate threads_pool;
use threads_pool as pool;
use std::thread;
use std::time::Duration;

let mut pool = pool::ThreadPool::build(4);

for id in 0..10 {
    pool.exec(|| {
        thread::sleep(Duration::from_secs(1));
        println!("thread {} has been waken after 1 seconds ... ", id);
    }, true);
}

pub fn execute<F: FnOnce() + Send + 'static>(
&self,
f: F
) -> Result<(), ExecutionError>

Similar to exec, yet this is the simplified version taking an immutable version of the pool. The job priority will be automatically evaluated based on the queue length, and for pool under pressure, we will try to balance the queue.

This method will be no-op if the pool is lazy-created (i.e. created via the build functions) and activate has not been called on it yet. If you are not sure if the pool has been activated or not, alter to use exec instead, which will initiate the pool if that’s not done yet.

Examples

extern crate threads_pool;
use threads_pool as pool;
use std::thread;
use std::time::Duration;

let pool = pool::ThreadPool::build(4);

for id in 0..10 {
    pool.execute(|| {
        thread::sleep(Duration::from_secs(1));
        println!("thread {} has been waken after 1 seconds ... ", id);
    });
}

pub fn sync_block<R, F>(&self, f: F) -> Result<R, ExecutionError> where
R: Send + 'static,
F: FnOnce() -> R + Send + 'static,